requirements.txt
azure_storage==0.37.0
Flask==3.0.3
Flask_Cors==4.0.1
openai==1.35.13
psycopg2==2.9.9
spotipy==2.24.0
